Ensure no two advisories represent the same issue (#182)
It dawns on me that the easiest way we might screw up the publication of an advisory is if we accidentally create a new JLSEC id that is representing the same alias or upstream advisory. This would be quite easy to do with multiple concurrent pull requests. We should probably tweak some toggles to make sure that CI is always run against a fresh merge with a queue or somesuch.